positive or negative

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -10,20 +10,20 @@
 int main(void)
 {
 	
-	long int n;
+	int n;
 	srand(time(0));
 	n = rand() - RAND_MAX/2;
-	if (n<0)
+	if(n<0)
 	{
-		printf("%ld is negative\n", n);
+		printf("%d is negative\n", n);
 	}
 	else if(n>0)
 	{
-		printf("%ld is positive\n", n);
+		printf("%d is positive\n", n);
 	}
 	else
 	{
-		printf("%ld is zero\n", n);
+		printf("%d is zero\n", n);
 	}
